requirement.txt
aiohttp==3.8.1
aiohttp-retry==2.8.3
aiosignal==1.2.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1
attrs==22.1.0
backcall==0.2.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docker==4.2.2
dpath==2.0.6
dulwich==0.20.45
dvc==2.24.0
dvc-data==0.4.0
dvc-gdrive==2.19.0
dvc-http==2.19.1
dvc-objects==0.2.0
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.10.0
entrypoints==0.4
executing==0.10.0
flatten-dict==0.4.2
flufl.lock==7.1
frozenlist==1.3.1
fsspec==2022.8.2
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
grandalf==0.6
greenlet==1.1.1
holiday==1.0.0
httplib2==0.20.4
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter_core==4.11.1
kombu==5.2.4
matplotlib==3.5.3
matplotlib-inline==0.1.3
missingno==0.4.2
mlflow==1.28.0
multidict==6.0.2
mysql-connector-python==8.0.30
nanotime==0.5.2
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.22.4
oauth2client==4.1.3
omegaconf==2.2.3
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pip==22.1.2
pipreqs==0.4.11
plotly==5.10.0
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
PyDrive2==1.14.0
pygit2==1.10.1
Pygments==2.12.0
pygtrie==2.5.0
PyJWT==2.4.0
Pympler==1.0.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pypiwin32==223
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.1
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.7.3
scmrepo==0.0.25
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==61.2.0
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.4.0
statsmodels==0.13.2
streamlit==1.12.0
tabulate==0.8.10
tomlkit==0.11.4
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.11
validators==0.20.0
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.1